pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hpcperfstats-tools"
version = "0.1.0"
description = "Command-line tools for querying an HPCPerfStats deployment"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "Texas Advanced Computing Center", email = "sharrell@tacc.utexas.edu" },
]
classifiers = [
  "Environment :: Console",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: GNU Lesser General Public License v2.1 (LGPLv2.1)",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
]
dependencies = [
  "requests",
  "python-dateutil",
  "pandas",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
]

[project.scripts]
hpcperfstats-jobstats = "hpcperfstats_tools.jobstats_cli:main"
hpcperfstats-sacct-gen = "hpcperfstats_tools.sacct_gen:main"

[tool.setuptools]
packages = ["hpcperfstats_tools"]

